fix(rocm): 8 fixes from 5th full-codebase audit (2 agents, line-by-line)
COMPILE ERRORS:
1. cuda::minimum: gpu_decode_rle.cu:190 uses ::cuda::minimum<uint32_t>
   (CCCL functor in namespace cuda, not cuda::std). Added cuda/functional
   shim with cuda::minimum<T>/cuda::maximum<T> + #include_next.

2. Telemetry headers: telemetry_context.hpp declares rust::Box<quent::Context>
   members; empty stub headers don't define these types. Created
   _types.hpp with minimal rust::Box<T>, uuid::UUID, and quent::* observer
   types. All 13 stub .rs.h headers now include _types.hpp.

3. __cccl_config.h: was a hard override (no #include_next), shadowing
   hipDF's real CCCL config. Added #include_next so hipDF's CCCL config
   takes precedence; minimal _CCCL_DEVICE macros are #ifndef-guarded fallbacks.

4. std::memcpy in __device__ code: cooperative_groups/memcpy_async.h used
   std::memcpy in __device__ functions. hip-clang may reject std::-qualified
   host functions in device context. Changed to __builtin_memcpy.

5. FULL_MASK: was uint32_t (0xFFFFFFFFu, 32 lanes) but warp_threads=64
   on AMD. Changed to unsigned long long (0xFFFFFFFFFFFFFFFFull, 64 lanes).
   Fixes FSST warp-reduce correctness on gfx942.

6. cuda/atomic,cmath,stream_ref,memory_resource,functional: were hard
   overrides shadowing hipDF's real CCCL. Rewrote all with
   #if __has_include_next → #include_next → #else fallback pattern.
   When hipDF is installed, its CCCL takes precedence.

CONFIGURE:
7. pixi.toml [feature.rocm]: added make + cxx-compiler (missing because
   no-default-feature=true prevented inheritance from top-level deps).

CLEANUP:
8. Removed dead _redirect.hpp (never included by any file; _fallback.hpp
   is the active redirect mechanism).

CMake if/endif: 50/50 balanced. Total shim files: 77.

+++ src/include/cuda/scan/detail/fsst.cuh
@@ -129,10 +129,9 @@ warp_compute_decomp_len(uint8_t const* __restrict__ comp_ptr,
 
 #pragma unroll
     for (uint32_t s = cub::detail::warp_threads / 2; s > 0; s /= 2) {
-      // HIP's __shfl_xor_sync requires a 64-bit mask (static_assert on ROCm);
-      // CUDA accepts both 32-bit and 64-bit. Cast to unsigned long long for
-      // portability — FULL_MASK (unsigned int) widens cleanly on both backends.
-      my_len += __shfl_xor_sync(static_cast<unsigned long long>(FULL_MASK), my_len, s);
+      // FULL_MASK is unsigned long long (64-bit) — satisfies HIP's static_assert
+      // for 64-bit masks. On NVIDIA the upper 32 bits are ignored (32-lane warp).
+      my_len += __shfl_xor_sync(FULL_MASK, my_len, s);
     }
     total += my_len;
